+++ backrooms/backrooms_builtins/brs_si_vector_load.py
@@ -14,4 +14,8 @@ SCRIPT = """
 
 
 def get_handler() -> StringHandler:
+    """
+    info: Gets script handler.
+    :return: StringHandler
+    """
     return StringHandler(NAME, SCRIPT)