+++ tests/test_files/__init__.py
@@ -1,5 +1,7 @@
 """
 Copyright 2021 Charles McMarrow
+
+Hold test files.
 """
 
 # built-in