+++ RELEASE_NOTES.rst
@@ -7,9 +7,11 @@ v1.0.0
 * Added more tests
 * Added more examples
 * Added start argument to Rule
+* Added StringReverse
 * Cleaned up code
 * Cleaned up logging
 * Fixed Typing
 * Fixed set_hallway_name None replacement bug
+* Fixed function stack corruption.
 * Made ThreadJoin to give up ThreadLock
 * Wrote more documentation